api/requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
pySmartDL==1.3.4
python-engineio==4.0.0
python-socketio==5.0.3
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.6.3
